teste_api_interface/backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
SQLAlchemy
pydantic
python-dotenv
pandas
